fix reference IP address && set .created_at date
The reference mpay24 IP address for the test mode was wrong, this
fixes it to the actual one.

Also set order.created_at when sending a new order to the payment
gateway. Before that .created_at was set when the order was first
generated -- which due to browser cookies could be weeks or months
before the actual order was send and marked as completed.

This also meant that all sorting within the admin interface was borken
(as .created_at was used for that). This fix at least allows for
more natural odering within the admin interface.