fix: Mpay24 confirmation IP verification should work behing nginx
Before this change we were just verifying the originating IP against
mpay24's IP which broke things behind the nginx reverse-proxy (as
nginx sets the originator IP to 127.0.0.1).

Now if we see an originator of 127.0.0.1 we check for the presence of
an HTTP_X_FORWARDED header and check against this.