WIP: start to bump spree version to 1.0 (or 1.1)
This started mostly with the namespace changes but while in there
I've started with refactoring most of the source code.

Now non-functional parts as XML generation and HTTP communication
are moved into special modules within the lib-directory. The incoming
response IP check has moved into the corresponding controller.

This will not work for now, but should be a better starting point than
before.